create a private shared crate!
