+++ Cargo.toml
@@ -28,6 +28,9 @@ debug = true
 
 [workspace.dependencies]
 bridge_types = { path = "bridge_types" }
+compile_state = { path = "compile_state" }
+sl-compiler = { path = "compiler" }
+slvm = { path = "vm" }
 regex = "1"
 lazy_static = "1"
