+++ compiler/Cargo.toml
@@ -8,9 +8,9 @@ edition = "2021"
 [dependencies]
 unicode-segmentation = "1.9"
 unicode_reader = "1"
-slvm = { path = "../vm" }
+slvm = { workspace = true }
 builtins = { path = "../builtins" }
-compile_state = { path = "../compile_state" }
+compile_state = { workspace = true }
 
 [build-dependencies]
 chrono = "0.4"
